pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pms_tiering_daemon"
version = "0.0.1"
authors = [
    { name="SimonBaier" },
]
description = "Plex media tiering daemon"
readme = "README.md"
license = { file="LICENSE" }
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "plexapi >= 4.12.1",
    "PyYAML >= 6.0",
    "prometheus-client >= 0.14.1"
]


[project.urls]
"Homepage" = "https://github.com/sbaier1/pms-tiering-daemon"
"Bug Tracker" = "https://github.com/sbaier1/pms-tiering-daemon/issues"